# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chat API",
    "version": "1.0.0",
    "description": "API documentation for auth, users, messages (conversations), and file upload/download.",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.yourapp.com/api",
      "description": "Production Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token in format: Bearer <token>"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": { 
            "type": "string", 
            "example": "64a1b2c3d4e5f6",
            "description": "Unique user identifier"
          },
          "name": { 
            "type": "string", 
            "example": "Saeed" 
          },
          "email": { 
            "type": "string", 
            "format": "email", 
            "example": "saeed@example.com" 
          },
          "avatarUrl": { 
            "type": "string", 
            "nullable": true,
            "example": "https://example.com/avatars/avatar.jpg" 
          },
          "createdAt": { 
            "type": "string", 
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "UsersList": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/User" }
          },
          "total": {
            "type": "integer",
            "example": 150,
            "description": "Total number of users"
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 50
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": { 
            "type": "string", 
            "example": "Error message" 
          },
          "error": {
            "type": "string",
            "example": "BAD_REQUEST"
          },
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "details": { 
            "type": "object",
            "nullable": true
          }
        }
      },
...
    },
    "parameters": {
...
      "userId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Unique user identifier"
      },
...
      "pageParam": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": { 
          "type": "integer", 
          "default": 1,
          "minimum": 1
        },
        "description": "Page number for pagination"
      },
      "limitParam": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": { 
          "type": "integer", 
          "default": 50,
          "minimum": 1,
          "maximum": 100
        },
        "description": "Number of items per page"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "message": "Unauthorized access",
              "error": "UNAUTHORIZED",
              "statusCode": 401
            }
          }
        }
      },
      "NotFoundError": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "message": "Resource not found",
              "error": "NOT_FOUND",
              "statusCode": 404
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "description": "Retrieve a paginated list of all users",
        "operationId": "getUsers",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/pageParam" },
          { "$ref": "#/components/parameters/limitParam" }
        ],
        "responses": {
          "200": {
            "description": "List of users retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsersList" }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/users/search": {
      "get": {
        "tags": ["Users"],
        "summary": "Search users",
        "description": "Search users by name, email, or phone number",
        "operationId": "searchUsers",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { 
              "type": "string",
              "minLength": 1
            },
            "description": "Search query (name / email / phone)",
            "example": "saeed"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": { 
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - query parameter missing",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "description": "Retrieve a single user by their unique identifier",
        "operationId": "getUserById",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/userId" }],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  }
}